requirements.txt
streamlit
moviepy==1.0.3
requests
python-dotenv
arabic-reshaper
python-bidi
google-generativeai